Major additions
Algorithm:
User inputs a manga name, a request is made to the appropriate manga site (selected by user; there is a default site called manganelo), the manga cover images from the user input appear. When the user clicks a cover image, a request will be made and the download will start.